implemented 2 enchantment logic
TODO: move apply effects method?

TODO: recode set_block effect? works for now...

TODO: add the rest of AdvancedEnchantments effects (have to find a way to make them simplified in config, format so far is easy to understand also have to change the example documentation in the config to match the new format.